в ООП запихнул player и управление
